Update and rename ImmortalUnlessBlocked.js to Immortal.js
